fixed: [AE] volume control values to use 0.0f - 1.0f range
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/AudioEngine@32457 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GUISound.h
@@ -34,7 +34,7 @@ public:
   void        Play();
   void        Stop();
   bool        IsPlaying();
-  void        SetVolume(int level);
+  void        SetVolume(float level);
   void        Wait(uint32_t millis = 500);
 
 private:
